Move AdvancedStatisticsSettingsFragment to a new file
