+++ AnkiDroid/src/main/java/com/ichi2/anki/preferences/AdvancedStatisticsSettingsFragment.kt
@@ -1,0 +1,32 @@
+/*
+ *  Copyright (c) 2022 Brayan Oliveira <brayandso.dev@gmail.com>
+ *
+ *  This program is free software; you can redistribute it and/or modify it under
+ *  the terms of the GNU General Public License as published by the Free Software
+ *  Foundation; either version 3 of the License, or (at your option) any later
+ *  version.
+ *
+ *  This program is distributed in the hope that it will be useful, but WITHOUT ANY
+ *  WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
+ *  PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with
+ *  this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.ichi2.anki.preferences
+
+import com.ichi2.anki.R
+import com.ichi2.preferences.SeekBarPreferenceCompat
+
+class AdvancedStatisticsSettingsFragment : SettingsFragment() {
+    override val preferenceResource: Int
+        get() = R.xml.preferences_advanced_statistics
+    override val analyticsScreenNameConstant: String
+        get() = "prefs.advanced_statistics"
+
+    override fun initSubscreen() {
+        // Precision of computation
+        requirePreference<SeekBarPreferenceCompat>(R.string.pref_computation_precision_key)
+            .setFormattedSummary(R.string.pref_summary_percentage)
+    }
+}
